Move pre2015 IPAC4 code back into upac.h and ipac.c files.  Added IPAC4 to lookup table functionality.
