+++ src/libs/ipacseries.c
@@ -121,6 +121,7 @@ convertIPACKey (enum ipac_boards_t bid, json_object* jobj)
   switch (bid)
   {
     case PRE_IPAC2_BOARD:
+    case PRE_IPAC4_BOARD:
       return convertIPAC(jobj);
       break;
 
@@ -752,9 +753,9 @@ int keyLookupTable[5][60] = {
  -1, -1, -1, 16, 20, 29, 30, 18, 22, 31, 32, -1, -1, -1, -1},
 
 /* Pre2015 IPAC4 */
-{1, 6, 2, 4, 13, 14, 9, 11, 100, 105, 101, 103, 112, 113, 108, 110, 3, 8, 5, 10, 7, 12,
- 24, 26, 15, 17, 19, 21, 23, 25, 27, 28, 102, 107, 104, 109, 106, 111, 123, 125, 114,
- 116, 118, 120, 122, 124, 126, 127, 16, 20, -1, -1, 18, 22, -1, -1, 115, 119, 117, 121},
+{1, 6, 2, 4, 13, 14, 9, 11, 101, 106, 102, 104, 113, 114, 109, 111, 3, 8, 5, 10, 7, 12,
+ 24, 26, 15, 17, 19, 21, 23, 25, 27, 28, 103, 108, 105, 110, 107, 112, 124, 126, 115,
+ 117, 119, 121, 123, 125, 127, 128, 16, 20, -1, -1, 18, 22, -1, -1, 116, 120, 118, 122},
 
 /* Ultimate I/O */
 {5, 7, 1, 3, 28, 26, 32, 30, 17, 19, 23, 23, 31, 29, 10, 9, 8, 6, 4, 2, 24, 22, 20, 18,
